AJM fixes for Andrew's vector tests
This change puts us back into a world where everything builds again,
but we still need to transfer Andrew's test driver patches.

The main issue being fixed related to gcc being strict on converting
function pointers to 'void *' in an in-place new that did not strike
on the other Bloomberg platforms.

It may be that only gcc respected the rules for casting multiply-
indirect pointer types ('T***' etc.) where safe cv-conversions
on the intermediate levels of pointer nesting are *not* supported.