Add workaround for MSVC bug
The Microsoft compiler (so far tested with only VC2008) sometimes treats
a function pointer as a pointer-to-const, and at other times does not,
leading to an unfortunate incompatiblity when trying to deduce the base
'imp' type to dispatch the overloaded operator free functions.

The current workaround is enabled for all MSVC compilers but might be
more finely targetted if research shows that later compilers do not
have this bug.