Remove non-trivial test extensions
This patch is entirely about testing vector of function pointers,
so we draw the line at any extension of the current test driver to
handle vector-of-pointer cases beyond simply adding the pointer
case and seeing it pass.

We remove the speculative extension to the allocator-aware test as
pointer types are not allocator-aware, and while there is a missing
piece of test functionality regarding testing non-allocator aware
elements for test case 11, that requires some more focussed thought
and belongs in a larger patch to properly test vectors with a variety
of types using the bsltf machinery.

Similarly, the test pattern for range-construction does not handle
vectors of pointers as the test machinery assumes that it can slice
to a 'Vector_Imp<TYPE>' which is not true when 'TYPE' is a pointer.
This covers test cases 12 and 13.  It also highlights that test
case 17 for 'insert' has a similar issue.  Proper handling of these
test cases is deferred to a more extensive review.