Add Intel library building flags to options file.
The Intel library has multiple build modes.  For BDLDFP we have
elected to use the non-reference mode, with global flags and
context information.  The source files for the library need to be
compiled with this indicated.  Intel does this with extra '#define'
options.  The Intel library also generates many cast-qual warnings.
These warnings are now disabled.