Changes to flattening for portability.
Because the BDE build process has no notion of per-platform
package builds, we need a way to make sure that Intel DFP only
builds on supported platforms.  This commit adds a small wrapper
around the core C intel library files, which first checks with
BSLS_PLATFORM_* macros, as to whether inclusion of the Intel
code is suitable.