Simplify 'load' overloads
The open source version was missing the overload that could
handle three null pointers, as caught by the test driver.

In order to fix this, we took another look at the supplied
overloads and their reasons to exist, and produced a simpler
set that has the same effect.  Confirmed that the test driver
passes cleanly with this simpler set of overloads in place.