Fixed Decimal128 'operator=='.
The 'lhs' parameter was duplicated in the call to the underlying
'DecimalImpUtil::equal' function.  This caused erroneous results
in the test driver only for the 128-bit equality operator.  This
is fixed, and the 128-bit tests have been altered to print out
raw 128-bit values under comparison, on test failure, for human
analysis.