Changed 'platform' macros from bool to definition.
The 'bdldfp_decimalplatform' component defined a number of 1-or-0
macros which indicated presence or absence of compiler and library
features.  These have been changed into defined/undefined pairs,
to facilitate the more common idiomatic usage.  The macros are
defined to '1' when defined, to maintain backwards compatibility
with the original design.  (Undefined macros have a '0' value,
in preprocessor expressions.)