suppress unused expression result warning
In the bsls_unspecified.t.cpp test driver when compiled with clang, the
following would result with -Wunused-value being emitted:

  bool tmp = true;
  bool var = ((tmp == true), <some-unspecified-bool-object);

One solution is to have the left-hand side of the comma operator to have
a side-effect (e.g., assignment of tmp).  However, in this change, per
gcc' documentation for -Wunused-value, the left-hand side of the comma
operator is cast to 'void' to suppress the warning.