Safely reference IntelDFP sources in bdl+inteldfp
The waf and bde_build.pl build systems require that all source
and header files to be build must be in the root directory of the
package.  Intel's DFP library files used to be manually "flattened"
into the intel+dfp directory, using the 'flatten.sh' script.  To
facilitate easier building, and prepare for DFP release, these
flattened wrapper files have been added to the git repo.