Correct exported 'bsl' names for open source
This patch addresses a review comment in phabricator, and extends
it to fix the remaining 'bsl+bsl*.h' headers.  Any symbol that
does not *need* to be exported according to the C++03 standard is
guarded by a BDE_OMIT_INTERNAL_DEPRECATED macro, so they are not
exported in the open source distribution.  Identifiers are sorted
alphanumerically for expected conventience.