Initial porting work: IntelDFP to IBM and Solaris.
IntelDFP compiles out-of-the-box for Solaris, and some other
platforms.  When incorporated into the BDE build system, some
changes are necessary to make it compile.  Due to IBM's XLC only
supporting GCC extensions in C++, all of the builds will now build
the library in C++, not C.  A few minor patches to Intel's code
were necessary to facilitate building in C++.  (Malloc, and some
symbol issues.)