Fixed for IC beta
This set of changes is a mixed bag to support the IC beta branch.

First we remove the 'bde_verift' pragmas as they trigger an
unknown pragma warning on gcc, which we build with both
'Wall' and 'Werror'.

Secondly, we resolve issues around the deprecated-internal
constructor taking a single 'SharedPtrRep *' argument,
which creates all sorts of interesting ambiguities when
restored and interacting with the 'std' interface.

Finally, this branch includes some work-in-progress
working on improved documentation in repsonse to review
feedback.