Revert to an earlier solution for null pointers
This patch reverts to the idea that we can explicitly construct a
'shared_ptr' from 'NULL' or '0', but cannot implicitly create such
a shared pointer; this implies no assignment from NULL, no returning
NULL as a function result, and no passing 'NULL' as an argument to
a function taking a 'shared_ptr'.  It is believed that there are only
two hits in Robo (per IC beta) that would be impacted by this, so
that is the smallest set of changes to supply as overlays.

The open source BDE lacks the problematic constructor, so will work
correctly (and has corresponding tests in the test driver).  The
horrible hack to try to make even that above work on the internal
BDE is significantly less desirable than simply patching two files,
and removing an otherwise unused feature.