Test cases for DecimalImpUtil::convertToDecimalXX
The convert to decimal routines convert between various widths of
Decimal Floating Point numbers, such as 32->64, and 64->128.  The
testing needs to make sure that value is preserved across widening
conversion, usually when narrowing, and that when narrowing it
pins to 'Inf' or 'NaN' only when loss of precision would make an
approximation unrepresentable.