Silence a bad test in the 'intelwrapper'.
The 'intelimpwrapper' component wraps the intel headers, but it
seems that some system headers may be defining '_WCHAR_T_DEFINED',
even if we '#undef' it.  This is probably okay, since it is in
the reserved portion of the namespace for implementation names.
It may be that some implementations do not allow undefining this
symbol.