Fix 'numeric_limits' for Sun and IBM compilers.
The Sun and IBM compilers have trouble with the way that static
constants as members of fully-specialized class templates get
linked.  To faciliate the linker for these platforms, the member
constants of 'std::numeric_limits<DecimalXX>' specializations are
moved to a parent class which is partially specialized.  This may
still cause issues with the MicroSoft compiler.