Move 'Decimal' & 'DecimalImpUtil' tests from main.
The 'bdldfp_decimal.t' and 'bdldfp_decimalimputil.t' test drivers
took exceedingly long to build on the Solaris platform compilers.
This was because that compiler begins to struggle with inlining
functions when the body of the function containing the inlined
calls passes beyond a certain size.  By moving each test case to
a separate function, each of which is not specified as 'inline',
we prevent this behavior.