Add missing 'inline' keywords
Recent research for the flat-map project demonstrated that at least
some of our production compilers are sensitive to the use of the
'inline' keyword on function template definitions - despite the
prevailing wisdom that the compiler should do its own thing without
regard to the use or lack of such hints.

This patch applies 'inline' to any function in the bslstl_string,h
heeader that has no more than 2 or 3 expressions, which hereby is
arbitrarily deemed a likely good candidate for inlining.  BSLS
ASSERT expressions were discounted when making this determinateion.