Fix gcc 4.8 warnings in test driver
This patch fixes a number of warnings about converting integers of
the wrong size directly to pointers.  In several cases, the test
logic was not quite what was intended, but would lead to a benign
pass anyway.

It is worth noting that in applying the specific form of fix taken
here, we are taking a short-cut that we know the test driver does
not adequately test 'vector' holding awkward types.  That should
be the basis of another ticket.