Fix allocate_ and make_shared
There is a divergance between the bslma::sharedptrinplacerep type and the blstl
equivalent that supports generic C++11 allocators.  The allocate_ and make_shared
functions are all written expecting the semantics of the latter, which stores a
bsls::ObjectBuffer of the desired element type, which is separately initialized
after the rep object itself has been allocated and constructed.  This second
phase of initialization correctly passes along allocators to the contained
object if it supports the given allocator type.  Conversely, the bslma componets
are passed all the arguments at construction, and do NOT pass on the blsma
allocator.  This is the existing behavior of the createInplace static member
function, and we do not want to change that behavior in this patch.

The combination of the two means that the (broken) behavior to be fixed was
the allocate_ and make_shared methods default constructing a data member, and
then overwriting it with a subsequent in-place construction, without first
destroying the default object.  Naturally, this would fail to compile if the
element type were not default constructible.

The quick patch is to forward all such calls to the correctly implemented
allocate_shared with a C++11 allocator which wraps the bslma allocator.  We
may want to revist this design using an updated bslma rep component again
in the future.