bsls_platform: Updated Clang GCC version selection
Clang, by default, defines pre-processor macros that indicate the
compiler is GCC and that it is specifically version 4.2.1.  BDE
already previously overrode this selection to indicate version
4.4.0 because Clang supports features found in the latest version(s)
of GCC.  This change makes Clang >= 3.3 mimic GCC 4.8.1, as both
compilers were the first versions to be C++ compliant in terms of
compiler features.  This will make Clang >= 3.3 compilation correctly
include GCC atomic intrinsics that are present only in GCC >= 4.7.