Remove long-deprecated (incomplete) specialization
For a while after we switched the semantics of containers keyed
by 'const char *' to use the pointer value rather than the string
value, we declared but never defined a template specailization for
'bsl::hash<const char *>'.  This was supposed to have been removed
around BDE 2.20 but appears to be present in the latest version
supporting modular hashing.  This patch finally removess the
deprecated-internal code specifically designed to break hashes
for this specific pointer type.