Audit and cleanup recursive epilog
This makes sure that the expected shared pointer features are exposed
through <bsl_memory.h>, including base classes used for user customization
of their own shared pointers.

We also audit the other headers, which look to be in good shape, and
inline the previous <bslstl_sstream.h> header which is an unnecessary
addition to the 'bslstl' package.  If the include-multiple-headers
without recursion idiom becomes important, it should be implemented
in 'bsl+bslhdrs' rather than by adding yet more components to bslstl.