Third-party builds for inteldfp and decNumber
Migrated inteldfp (The Intel Decimal Floating Point library) and
decNumber from the 'bdl' package group into the third-party
directory. Created waf 'wscript' files to for building 'decNumber' and
'inteldfp' independently and within the bsl repository -- decNumber does
not come with a build system. inteldfp has a makefile which is difficult
to port to multiple platforms.

Note that the the third-party directory contains the unaltered version
of decNumber library. Also note that the initial implementation was done
by Adam Martin, with improvements and fixes by Chen He.