Fix bslstl::basic_stringbuf str() assignment
Previously, when assigning a string to a stringbuf via the 'str()'
method, the get area was updated before the put area.  As the location
of the egptr is based off a calculation which involves the put area,
it seems like the put area should be updated first.  This commit makes
that change.

Signed-off-by: Daryl Haresign <github@daryl.haresign.com>