long double aligned at 8-byte ok on Linux on POWER
IBM's response to PMR 04669,7TD,000 confirms that 8-byte alignment
returned from standard malloc() is fine for long double on Linux
on POWER.

This was the response from the IBM ToolChain team:

    The "IBM long double" is implemented as a pair of double floats. The
    default behavior is to use normal load float double / store float
    double operations in pairs.

    As such double word alignment is sufficient for IBM long double. But
    for optimal load/store performance we prefer quadword alignment.

    If you are concerned about the alignment from libc malloc you can
    try the tcmalloc included in the AT mcore-libc RPM. tcmalloc will
    return quadword aligned storage for both -m32 and -m64.

IMPLEMENTATION NOTE:
bsls_alignmenttotype.h makes the assumption that a type must
be aligned to a boundary as large as its size.  (Aside: given
this assumption, one may wonder why there is all the template
metaprogramming complexity in bsls_alignment* components.)  The
assumption of matching alignment and size for native types is
incorrect for 'long double' on Linux on POWER (see above):
    PRIORITY = sizeof(AlignmentImpMatch::match(Tag(),
                                               Tag(),
                                               MaxPriority()))
since templates for AlignmentImpMatch::match() are defined using
    #define BSLS_ALIGNMENTIMP_MATCH_FUNC(T, P)                         \
      bsls::AlignmentImpTag<P> match(bsls::AlignmentImpCalc<T>::Tag,   \
                                     bsls::AlignmentImpTag<sizeof(T)>, \
                                     bsls::AlignmentImp_Priority<P>)
where the alignment for 'long double' can be 8 when using malloc(),
but the sizeof 'long double' is 16.  This does not appear to be a
problem for the use in bsls_alignedbuffer.h.  Still care should be
taken if using bsls_alignmentotype.h with 'long double', and for that
matter, when using 'long double' in general.