Fix initial sharedptr bugs in C++11 builds
This fixes the variadic forms of 'allocate_shared' and 'make_shared' to
use the same idiom as the more widely tested and deployed C++03 version.
This appears to be an oversight where development updated how the C++03
overloads would be implemented, but did not reflect that back to the
C++11 code line.

However, code-inspection shows that this new path is buggy in a different
way that we are not currently testing for.  Specifically, if the constructor
of the new element throws, then the destructor will be run for the object
that was never created.  Next change-set should be a patch to the test
driver that exposes this bug.