Add missing test concern to make_shared tests
A subtle error was revealed by code inspection where a destructor would run
for an object that threw an exception aborting its constructor when using
one of the allocate_shared or make_shared overloads.  This patch adds a simple
extension to the end of each test case to expose the issue.  Now that the bug
is exposed by the test drivers, we need to fit it.