Adapt forwarding type for MSVC compiler bug
MSVC cannot forward arrays of unknown bound by reference, as it treats
references to such types as if they were arrays of 0-length.  This bug
is fixed in MSVC 2012.  For earlier compilers, we omit that template
specialization, and simply drop support for forwarding a type that the
compiler would not allow to be forwarded through non-generic code either.