Disable all tests for array of unknown bound on broken compilers
Disable every test case for array-of-unknown-bound for compilers that
are known to have bugs that are extremely difficult, or, more likely,
impossible to work around.  This includes Sun CC, IBM xlC, and older
releases of Visual C++.

An added workaround notices that patches for the Sun compiler were not
complete in the last change set, as the compiler was actually tripping
on some internal compiler error that was reported in a manner that did
not jump out as easily when grokking the error/warning messages.  For
the moment, we disable *all* testing of arrays on Sun.  This is not a
desirable state of affairs, but we want to see a clean build for all
the other tests and platforms tonight before trying to track down the
exact cause of breakage in the Sun compiler.