Fix third party dpkg build
In the dpkg build process for bsl-internal, debian/control file is
generated from debian/control.in. The previous patch updated
debian/control instead of debian/control.in.  I believe debian/control
can be removed, but I will leave the file as is because we are very
close to a new release.