Improve the performance of decimal64FromDouble
The methods to convert binary floating point to decimal are slower than
necessary due to the use of streams.  According to Dietmar Kuhl, the
steams being used needs to use 2 locale objects, which each require at
least one synchronized increment.  Changing the implementation to use
'snprintf' improved the performance by a factor of 4.