Workarounds for testing with Sun CC
This is the final (planned) set of platform-specific test driver
workarounds, in this case for the Sun CC platform.  As the patch
is all about making the test driver pass a compile, for reasons
unrelated to the (untouched) component header, this seems to be
a reasonable way to pass the tests.