Fix the destroy-a-non-constucted object bug
The bug in both 'allocate_shared' and 'make_shared' is that the
proctor used to guard the in-place representation after intial
allocation releases that rep by reducing the ref-count, which
reaches zero and runs the destructor before disposing of the
memory.  This means running the destructor for the in-place
element that was never constructed.  The fix is to instead call
'disposeRep' which simply deallocates the memory for the
rep without touching or destroying the ref-count memebers.
Given the very sepcific use pattern that this proctor is
guarding, where the reference count is either 0 or 1, and
the knowledge that the atomic counters have trivial
destructors, this simple fix appears to also be the correct
one.