Updated in response to review comments
Addressed all issues from the phabricator review, and a couple more.
Finally, note that there is one corner that this patch does not yet
resolve.  If the user calls 'allocate_shared' with a custom C++11
compliant allocator that has non-trivial state, then a copy will be
made but its destructor will not run.  The fix for this corner is a
rather more substantial change than simply tweaking the proctor for
the inplace rep, essentiallly requiring that we do call the rep
destructor without destroying the in-place object (that fails to
construct) as before - and this will mean adding appropriate
forwarding constructors to the inplace rep object itself, among
other things.  We deliberately defer that final fix until we can
assume a C++11 compiler from all of our clients, and so rely on
perfect forwarding with a single variadic signature.  This corner
is not something we see as a critical blocker in the meantime.