Patch a couple of issues in the nightly build
gcc 4.1.2 is another platform that has problems with references to
arrays of unknown bound.

If we are going to call 'native_std::move' in C++11 build modes,
then we also need to #include the <utility> header.  Ideally we
would provide our own implementation of 'move' in 'bsls_util'
in order to avoid the risk of introducing cycles in the 'bsl+*'
levels.