Remove xlC 10.x workaround from associative containers
Resolving a 3 year old DRQS ticket to remove a workaround we applied, pending
the arrival of an updated xlC compiler.  The 11.2 compiler has been the base
compiler in production for most of that time, so finally make the change.
This moves the base requirement for BDE/bsl support up to xlC 11.2 from 10.1,
and the open source docs should be updated accordingly when this patch is
merged.

The issue is only working around BDE coding rules that methods are never
defined in the class, but always as out-of-line (potentially 'inline')
functions at the end of header, or otherwise in the .cpp.

As part of this cleanup, fixed any obvious issues reported by bde_verify
but all headers are still around 30 warnings from being bde_verify clean.
As the test drivers were not touched, they remain very bde_verify hostile.