Fix  MSVC warnings in bslstl_hashtable.h
The three tests that a pointer assignment is (not) null are producing
over 1000 warnings from this header file when building the test driver
on Windows.  That is a good sign that this really is a nuisance for
Windows customers using the unordered containers (actually, at least
5 test drivers are involved here, hash table and the 4 unordered
containers, but that is still a strong indication of many redundant
warnings).

The recommended MSVC code pattern to silence the warning is:
  if(0 != (assignment))
as the simpler alternative for gcc does not solve the problem:
  if ((assignment))