Fix untested branches in allocator traits test driver
MSVC 2013 unreachable code warnings highlighted a problem with the macros
used to define test case 8 - they include a 'break' as if part of a 'switch'
statement, so the first use of such a macro jumps out of the block, skipping
the rest.  AFAICT, this code has not actually been tested since the driver
was first written.