Fix seeded hash test driver to avoid user defined literals
This test driver uses string literal concatenation with a macro that
expands to the appropriate printf format flag on different platforms.
There needs to be whitespace between the macro name and the preceding
string literal it is concatenating with, or it parses as a user
defined literal in C++11 or later.