First pass at fixing typos in headers
This pass deliberately avoid 'bslstl' as it would be risk
clashing with other merges due to hit that package in the
near future.  Otherwise, this fixes most of the 'real'
spelling errors detected by 'bde_verify' in the headers
of packages in the 'bsl' package group.  It also fixes
any trailing whitespace issues that are reported no the
same set of components.