Merge workarounds from a seconds traits-fixing branch
There are two branches racing to land fixes to several type traits,
and due to using slightly different techniques, they merge (mostly)
cleanly into the integration branch, but cause ambiguities by adding
both solutions (that should be unique).

This attempts to consolidate a single version of the fixes for the
is_enum and is_convertible trait, that are exercised by the test
driver of forwarding type.