Additional traits tests to track IBM bug
The 'is_enum<int const [4][2]>' test is failing as IBM is, for
some as-yet unknown reasons, recursively re-instantiating the
same template rather than removing the 'const' qualifier when
trying the template expansion that 'is_enun<const T>' is the
same as 'is_enum<T>'.

This patch enhances the 'is_const' and 'is_volatile' test
drivers to test a broader variety of types, including arrays
of higher rank, and so demonstrates that simply detecting the
cv-qualifier is not the problem here.

The 'is_enum' is similarly enhanced to better detect volatile
qualifier issues on the same test cases, and to confirm that
arrays of enumerations are not reported, in turn, as eunms.