Correct handling of cv-qualifiers in IBM type traits
This patch tests for and corrects some issues with the
handling of cv-qualifiers on arrays of more than one
dimension with the IBM xlC compiler.  This patch is
intended to be sufficient to land the ongoing sequence
of commits for the 'bslmf_forwardingtype' brach, without
being an extensive validation of the affected traits, or
across all type traits.  That should be handled by a
subsequent patch series.