Workaround for gcc 4.7 and 4.8
It appears gcc's behavior wrt overload resolution and rvalue/lvalue
references went through an awkward middle-age around the 4.7 and
4.8 release cycle, that caused an ambiguity on several of our test
cases.  Given the intended usage of this compenent, the simplest
workaround is to simply remove the lvalue overload of 'move' for
the case where rvalue references  (and forwarding references for
perfect forwarding) are available, without specifically testing
the compiler version.