Update function pointer traits doc for varargs
Attempt to provide correct documentation regarding functions
taking a C-style varargs parameter.  Considered updating the
non-varags function traits doc at the same to suggest they
require functions that take 'exactly' N arguments, but that
would not be correct for functions with default arguments.
This is one of those components where attempting the specify
precisely the effect of the code in English is generally more
open to misinterpretation than simple reading the code.