Update forwarding-type usage example to be more relevant
The original usage example induced unnecessary copies, by breaking
down the steps and usage of the forwarding type usage while
illustrating and documenting usage.  While this may have given a
slightly clearer example of the finer details of the process, it
was not a very good motivation for the component itself, which is
the purpose of a usage example.  Example updated accordingly, as
recommended by Hyman in review.