Remove BDE_OMIT_INTERNAL_DEPRECATED define on Windows
For testing purposes, and earlier Windows devkits, we defined
BDE_OMIT_INTERNAL_DEPRECATED in order to emulate building against
the open source distribution, and to keep Windows clients relatively
clean.  That no longer works for the WAF-built distributions we are
now deploying for the devkits as of the BDE 2.23 release.  There are
specific incompatibilities in the a-ossl package (which we have not
yet established a clear owner who would be responsible for deploying
a fix, although that is likely to ultimately be us again).  However,
in general we cannot assume that BAS with move in lock-step with
BDE choosing to deprecate something, and cannot unilaterally make
that choice for all Windows consumers.  Instead, Windows clients
should be free to make their own choice regarding defining that macro
for their own cleanliness, just like our Unix clients.  That still
means we need to push a fix to 'a-ossl' at some point using std
shared_ptr APIs, specifically '.get' rather than '.ptr' to access
the held pointer.