Fix bsltf::AllocTestType test driver on C++11
The latent problem is that the component relies on triggering an
assertion in an object's destructor to indicate that it was
bitwise-moved when it should not have been.  Our usual testing
technique works fine in C++03, but fails in C++11 when the
destructor becomes implicitly 'noexcept', so throwing the test
exception type out of the assert handler triggers a call to
'terminate' for violating the exception specification.

This patch provides a change that will work in both C++03 and
C++11.  We create a custom assertion handler for just this
test, that indicates success by calling 'exit(0)' and so not
needing to return control to the test case itself.  There are
a few additional tweaks to make sure that we do not lose any
already reported test failures, but that is the principle.

Additionally, fixed any gross issues reported by bde_verify,
without touching the component header or cpp files.  This
test driver is still not bde_verify clean, but is much
closer.