Fix is_cv-qual traits for MSVC
The Visual C++ compiler does not correctly handle arrays of
cv-qualified elements in partial template specialization, and so
gives the wrong result for 'is_const<const int[10]>::value', for
exmaple.

Having confirmed that arrays of cv-qualified elements are correctly
dispatched to functions overloads on both 'T&' and 'const T&', we
choose to fix the trait, rather than the test driver, concluding
that the compiler bug is inconsistent so that there is no way to
be true to the compiler anyway.  Therefore, it is better to have a
consistent trait, that agrees with the C++ Standard, and what a
conformant compiler should do in all, rather than some, cases.