Replace iostream with printf in atomics test driver
The Clang compiler/library complains about an ambiguous 'atomic_int' in
this test driver, as (presumably) Clang <iostream> is using atomics
for efficiency somewhere under the covers.  'atomic_int' is a standard
type, exposed by the 'using namespace std' at the top od this test
driver.

As we are levelized below 'bsl+stdhdrs', the correct solution is to
replace legacy test code with the <bsls_bsltestutil.h> support -
only it turns out that this platform-specific support component
has no actual test cases to run.  There is a single -ve numbered
performance test case, and a few messages printed from that test
case using 'cout'.  This is trivially replaced with the platform
runtime 'printf' instead, avoiding the accidental collision with
the standard name.