Final cleanup before submitting for review
This patch simply eliminates warning suppression for unused
identifiers that are actually used in the test driver, and
makes the #include pattern consistent, including whitespace,
across the package.  Attention is paid as, if this passes
review, it should be the model for all other test drivers
in the 'bsl' package group.