is_enum test driver workaround for MSVC
The metaprogramming to determine whether something is an
enum involves trying to deduce a reference to to that
type in a function call, and MSVC has a bug forcing a
hard (non-SFINAEable) error for arrays of unknown bound
in such circumstances.  We simply disable those tests
on older MSVC compoilers for now.