Disable variadic templates for MSVC 2013
The 'bslmf_typetraits' component demonstrates sufficient problems with
MSVC variadic template support that we defer the feature-detection
macro to the 2015 compiler, that in early testing seems to work fine.

I looked into adding the apparent failures from the typelist component
to the bsls_compilerfeatures test driver, but the cut-down form of the
test is already present, and I don't think it is worth the time chasing
down larger, more specific failure to include in this test driver, or
it will become exceedingly large for minimal gain.