Add header for 'int32_t' typedef
A portable header for the C++11 typedefs in <cstdint> has just
been added, <bsl_cstdint.h>, that will give a correct answer
even for platforms where the native header is not available.
Use that to fix the missing header problem, rather than fix
the code to not use the typedef.