Ensure memory written by allocator_traits test driver
The destructor for one of our test types overwrites the
value of an integral data member.  gcc, when building
as 'opt', notes that it is not possible to inspect that
value after the object is destroyed, so omits the
write to the integer.  This breaks tests that look to
examine the remains of destroyed objects to verify that
the destructor ran correctly.

The simplest fix is to make that one data member
'volatile'.  In addition, provide a well-defined way
to get the address of that data member before the
test object is destroyed.