bslstl::StringRef: always call width(0) in op<<
Correct bslstl::StringRef's operator<< to always call stream.width(0),
in line with the behavior of bsl::string's operator<<.

This fixes the output of the following code:

    bsl::cout << bsl::setw(8) << bslstl::StringRef("12")
                              << "34"
              << "\n";

which prior to this commit would ouput "      12      34".  After this
commit, it ouputs "      1234", which is the same output if the 12 were
printed as either a const char* or a bsl::string, rather than a
bslstl::StringRef.