Correct initializer_list header for gcc prior to 4.7
For gcc prior to 4.7, the <initializer_list> header would be #included by
standard headers for containers, even in C++03 builds.   This would then
break in our BSL_OVERRIDES_STD mode builds, where the attempt of the
intercept header to import 'std::intializer_list' into namespace 'bsl'
would fail to compile.

Testing this in C++11 mode to confirm that nothing accidentally breaks
turns up a number of issues in test drivers that still need to be
addressed by a subsequent patch - although it looks like we are good
to go for gcc 4.6 and later.