bdlb_guid: adds 'hashAppend' overload.
This commit introduces a straightforward overload of the 'hashAppend' free
function.  The implementation of the function passes the data of the guid
object to the hash algorithm.

This commit adds a test case to verify that guid objects can be hashed by the
standard 'bsl::hash<bdlb::Guid>' as well as 'bslh::Hash<>', and verifies that
hashes produced are distinct for distinct guid objects.