More cleanup, removing 'bindA'
This patch removes 'bindA' from 'bdlf_bind', and cleans up
most of the associated doc.  The rest of the repository is
updated to allow for the fact this function no longer exists.

Additional cleanup was performed on the 'bdlf' package to
remove use of deprecated 'bslalg' traits facilities, in
preference to using the 'bslmf' and 'bslma' traits directly.