Implement std p0033r1
Improve testing of 'bsl::enable_shared_from_this' and
implement the standard proposal p0033r1, which has two
aspects.  First, it adds a 'weak_from_this' overload,
then it assumes that 'enable_shared_from_this' objects
always have a valid (but potentially expired) 'weak_ptr'
member, and reset this only for the first 'shared_ptr'
constructor that tries to take ownership of the object.
Subsequent 'shared_ptr' constructors, such as when
passing a pointer with a no-op deleter, do not establish
true ownership, and so should /not/ rebind the 'weak_ptr'.