Additional testing for P0033R1
This patch completes the support for well-defined
behavior of 'enable_shared_from_this' in almost all
imagined circumstances, per the standard proposal
P0033R1 for C++17.  These revised tests caught a
variety of subtle bugs, notably in the iteraction
of aliased managed pointers, and implicit accesses
to the dead object when the inital 'shared_ptr' to
bind the 'd_weakThis' member is NOT a true owning
'shared_ptr', i.e., it has a no-op deleter or
similar.