Provide clearer constructor for 'weak_ptr::lock'
'weak_ptr::lock' creates a 'shared_ptr' object in a particularly
subtle way, which includes calling a constructor that is documented
as being a general purpose constructor (for clients implementing
their own shared rep types) that is not otherwise easily called,
due to the interaction of a more general constructor template
stealing such calls in overload resolution unless the parameter
types are /exact/ matches, and the typical use case is that at
least one of the two pointer types involved is to a derived
class.

This patch separates out the special case of locking a weak pointer
into its own 'shared_ptr' constructor, which is selected by passing
an extra 'tag' parameter.  This allows us to distinguish whether or
not the special dance for 'enable_shared_from_this' derived types
should happen.  The pre-existing constructor would be required to
perform that dance when called from 'bcec_sharedojectpool' for
example, yet that dance must NOT happen when locking a weak pointer
to avoid accessing a destroyed object in some very obscure
circumstances.

Getting the test case right showed just how subtle and dangerous
this constructor is, and we might consider making private and
'weak_ptr' a friend of 'shared_ptr', although that would make it
effectively impossible for users to create their own custom weak
pointer types.  Currently we have no motivating examples for such
support, although we know that custom shared pointer rep types
are used.